add additional fdatasync before close() on writes
