+++ pkg/disk/fdatasync_linux.go
@@ -1,0 +1,38 @@
+// +build linux
+
+/*
+ * Minio Cloud Storage, (C) 2020 Minio, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package disk
+
+import (
+	"os"
+	"syscall"
+)
+
+// Fdatasync - fdatasync() is similar to fsync(), but does not flush modified metadata
+// unless that metadata is needed in order to allow a subsequent data retrieval
+// to  be  correctly  handled.   For example, changes to st_atime or st_mtime
+// (respectively, time of last access and time of last modification; see inode(7))
+// do not require flushing because they are not necessary for a subsequent data
+// read to be handled correctly. On the other hand, a change to the file size
+// (st_size, as made by say ftruncate(2)), would require a metadata flush.
+//
+// The aim of fdatasync() is to reduce disk activity for applications that
+// do not require all metadata to be synchronized with the disk.
+func Fdatasync(f *os.File) error {
+	return syscall.Fdatasync(int(f.Fd()))
+}
